Apply sort in filters; move SortTypew from /mock to /data
